pyproject.toml
[project]
name = "prov"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
]
